создает data.js. main.js - делате главным модулем
